Add files via upload
This Python script is a tool for downloading YouTube videos with Pytube. It allows the user to enter a YouTube video URL and downloads the video in the highest resolution available. It also adds a suffix to the filename based on the last 3 characters of the URL. The script saves the video as a .mp4 file by default, but it can be changed to a .mov file by uncommenting the alternative line of code. The script prints a message when the download is completed.